Non-working but fairly complete camera setup
I made some change that is causing the splitter output buffer pool
allocation to fail. This happens regardless of which output's pool I try
to allocate first.

Previously I was seeing reasonable buffers passed to the raw splitter
output callback, but the encoder callback was getting buffers of 67 B
(way too small). Not sure what I did to break it, but either way it
doesn't work.

Seems like a good time for a checkpoint.