Sensor is sending messages (metadata only)
Added toml11 submodule--eventually the capture application should read
in the server host and other config options from a file. The receiver is
using toml, so let's use it here too.

Pi application can now establish a connection to the server and send
image metadata. The next step is sending the image data, too.