Cleanup in prototyping processing script, w/ notes
See the notes in the script for where things are. Generally speaking,
identifying connected components in the image is working. What needs
work is filtering out regions that are too large--these are caused by
bright spots in the image. I attempted to address this by blurring the
image and using that as a density image, but the edges of the blurred
regions are still not filtered out. Too-small regions are also a
problem, but easy enough to remove.

The edge detection also deserves some consideration (i.e. order or
absolute value).